PA-Permit/requirements.txt
streamlit
crewai
crewai-tools
ollama
python-dotenv
pydantic
graphviz
pandas
httpx
litellm

